+++ scripts/baselines/import_contracts_baseline.json
@@ -70,9 +70,6 @@
         "aragora.security -> aragora.connectors",
         "aragora.security -> aragora.gateway",
         "aragora.security -> aragora.server",
-        "aragora.storage -> aragora.gauntlet",
-        "aragora.storage -> aragora.reasoning",
-        "aragora.storage -> aragora.server",
         "aragora.swarm -> aragora.connectors",
         "aragora.swarm -> aragora.mcp",
         "aragora.swarm -> aragora.server",
